Refactor Symbol class: Replace byte with int and clean up unused methods

+++ src/main/java/moba/server/repositories/Layout.java
@@ -63,8 +63,8 @@ public class Layout {
                         rs.getLong("YPos")
                     ),
                     new LayoutSymbol(
-                        rs.getLong("Id"),
-                        new Symbol(rs.getByte("Symbol"))
+                        rs.getInt("Id"),
+                        new Symbol(rs.getInt("Symbol"))
                     )
                 );
             }
